Project 11 challenge 3
Added a date field to the Book entity which is initialised when the Book is created (in the same way as the id field). The formatted date is handled using a computed property in the DetailView which also takes care of the fact that the date is an optional. That turns out to be quite handy because it means that it can provide a default date for any books that were added before the date field was added to the data model.